Stop using destructuring assign by now, as it breaks uglify-js
