+++ static/js/audio.js
@@ -1,6 +1,8 @@
-index = require('./index.js');
+index = require('./app.js');
 app = index.app;
-app.ports.setCurrentTime.subscribe(function([id, time]) {
+app.ports.setCurrentTime.subscribe(function(data) {
+	var id = data[0],
+		time = data[1];
     var audio = document.getElementById(id);
     audio.currentTime = time;
 });
